requirements.txt
click==6.7
eventlet==0.23.0
Flask==1.0.2
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
greenlet==0.4.13
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-engineio==3.8.2.post1
python-socketio==4.2.0
six==1.11.0
Werkzeug==0.14.1
WTForms==2.2.1
requests==2.22.0
